backup/requirements.txt
# Core Dependencies
langgraph>=0.1.15
pydantic>=2.6.0
redis>=5.0.0
chromadb>=0.4.24
aiohttp>=3.8.0
numpy>=1.26.0
docker>=7.0.0
psycopg>=3.1.0
asyncpg>=0.29.0
GitPython>=3.1.40
fastapi>=0.100.0
uvicorn>=0.25.0
streamlit>=1.30.0
ollama>=0.1.34
python-dotenv>=1.0.0
rich>=13.7.0
psutil>=5.9.0
python-json-logger>=2.0.7
watchdog>=3.0.0

# Machine Learning and Embeddings
sentence-transformers>=2.4.0
torch>=2.1.0
transformers>=4.38.0

# Security and Guardrails
PyJWT>=2.8.0
cryptography>=42.0.0

# Monitoring and Metrics
prometheus-client>=0.19.0
matplotlib>=3.8.0
pandas>=2.2.0

# Utilities
requests>=2.31.0
networkx>=3.2.0
PyYAML>=6.0.1
tenacity>=8.2.0
humanize>=4.9.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
coverage>=7.4.0
black>=24.1.0
mypy>=1.8.0
ruff>=0.2.0

transformers
torch
accelerate